# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: NoteHub API
  version: "0.1.0"
  description: |
    NoteHub REST API および WebSocket 仕様。

    ## 認証
    - Access Token: `Authorization: Bearer <token>` ヘッダー（15分有効）
    - Refresh Token: HttpOnly Cookie `notehub_refresh_token`（`/api/auth` パス）
    - CSRF: Cookie `notehub_csrf_token` と Header `X-CSRF-Token` の Double Submit（`POST /api/auth/refresh`, `POST /api/auth/logout`）

    ## レスポンス形式
    成功時は `{ "data": ... }`、失敗時は `{ "error": { "code", "message", "requestId?" } }`。
servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://localhost:8082
    description: Local integration test
...
paths:
...
  /api/workspaces/{workspaceId}/members/{userId}/suspend:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/UserId"
    post:
      tags: [Account]
      summary: アカウント停止（ホストのみ）
      operationId: suspendAccount
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SuspendAccountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/workspaces/{workspaceId}/members/{userId}/reactivate:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/UserId"
    post:
      tags: [Account]
      summary: アカウント復帰（ホストのみ）
      operationId: reactivateAccount
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ReactivateAccountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/workspaces/{workspaceId}/members/{userId}/delete-account:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - $ref: "#/components/parameters/UserId"
    post:
      tags: [Account]
      summary: アカウント論理削除（ホストのみ）
      operationId: deleteAccount
      description: 仕様書では `/delete` パスだが、実装は `/delete-account`。
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DeleteAccountResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/workspaces/{workspaceId}/documents:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [Document]
      summary: ドキュメント一覧
      operationId: listDocuments
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/DocumentListItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
    post:
      tags: [Document]
      summary: ドキュメント作成
      operationId: createDocument
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/CreateDocumentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /api/documents/{documentId}:
    parameters:
      - $ref: "#/components/parameters/DocumentId"
    get:
      tags: [Document]
      summary: ドキュメント取得
      operationId: getDocument
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/GetDocumentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Document]
      summary: ドキュメントタイトル更新
      operationId: updateDocument
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDocumentRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/UpdateDocumentResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    delete:
      tags: [Document]
      summary: ドキュメント削除
      operationId: deleteDocument
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DeleteDocumentResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/WorkspaceForbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access Token（15分有効、メモリ保持）
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: CSRF Cookie `notehub_csrf_token` と同値
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    TooManyRequests:
      description: Too Many Requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    WorkspaceForbidden:
      description: Workspace access or permission denied
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    ApiResponse:
      type: object
      required: [data]
      properties:
        data: {}
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: 入力値が不正です
            requestId:
              type: string
              format: uuid
...
    SuspendAccountResponse:
      type: object
      required: [userId, status, suspendedAt]
      properties:
        userId:
          type: string
          format: uuid
        status:
          type: string
          example: suspended
        suspendedAt:
          type: string
          format: date-time
    ReactivateAccountResponse:
      type: object
      required: [userId, status]
      properties:
        userId:
          type: string
          format: uuid
        status:
          type: string
          example: active
    DeleteAccountResponse:
      type: object
      required: [userId, status, deletedAt]
      properties:
        userId:
          type: string
          format: uuid
        status:
          type: string
          example: deleted
        deletedAt:
          type: string
          format: date-time
    CreateDocumentRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 100
        content:
          type: string
          maxLength: 524288
    CreateDocumentResponse:
      type: object
      required: [id, workspaceId, title, content, createdBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        createdBy:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    DocumentListItem:
      type: object
      required: [id, title, updatedBy, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        updatedBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
    GetDocumentResponse:
      type: object
      required: [id, workspaceId, title, content, updatedBy, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        workspaceId:
          type: string
          format: uuid
        title:
          type: string
        content:
          type: string
        updatedBy:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
    UpdateDocumentRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 100
    UpdateDocumentResponse:
      type: object
      required: [id, title, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        updatedAt:
          type: string
          format: date-time
    DeleteDocumentResponse:
      type: object
      required: [documentId, deletedAt]
      properties:
        documentId:
          type: string
          format: uuid
        deletedAt:
          type: string
          format: date-time
...